@@ -13,19 +13,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.0.0",
-    "@angular/compiler": "4.0.0",
-    "@angular/core": "4.0.0",
-    "@angular/forms": "4.0.0",
-    "@angular/http": "4.0.0",
-    "@angular/platform-browser": "4.0.0",
-    "@angular/platform-browser-dynamic": "4.0.0",
-    "@angular/router": "4.0.0",
+    "@angular/animations": "4.0.0",
+    "@angular/common": "4.1.3",
+    "@angular/compiler": "4.1.3",
+    "@angular/core": "4.1.3",
+    "@angular/fire": "^5.1.0",
+    "@angular/forms": "4.1.3",
+    "@angular/http": "4.1.3",
+    "@angular/platform-browser": "4.1.3",
+    "@angular/platform-browser-dynamic": "4.1.3",
+    "@angular/router": "4.1.3",
     "admin-lte": "2.3.11",
+    "angularfire2": "^5.1.0",
     "core-js": "2.4.1",
+    "firebase": "^4.13.1",
     "font-awesome": "4.7.0",
     "intl": "1.2.5",
     "jquery": "3.1.1",
+    "ng2-currency-mask": "^5.3.1",
+    "ngx-toastr": "6.0.0",
     "reflect-metadata": "0.1.9",
     "rxjs": "5.1.0",
     "ts-helpers": "1.1.2",
@@ -33,7 +39,7 @@
   },
   "devDependencies": {
     "@angular/cli": "1.2.7",
-    "@angular/compiler-cli": "4.0.0",
+    "@angular/compiler-cli": "4.1.3",
     "@types/jasmine": "2.5.38",
     "@types/node": "7.0.5",
     "codelyzer": "2.0.0",
